+++ WebAppMVC/Controllers/HomeController.cs
@@ -1,16 +1,25 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Data.Entity;
 using System.Linq;
-using System.Web;
 using System.Web.Mvc;
+using WebAppMVC.Models;
 
 namespace WebAppMVC.Controllers
 {
     public class HomeController : Controller
     {
+
+        private ApplicationDbContext _context { get; set; }
+
+        public HomeController()
+        {
+            _context = new ApplicationDbContext();
+        }
+
         public ActionResult Index()
         {
-            return View();
+            var allPosts = _context.Posts.Include(p => p.Utilisateur).ToList();
+
+            return View(allPosts);
         }
 
         public ActionResult About()
