+++ WebAppMVC/Controllers/ManageController.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNet.Identity;
 using Microsoft.AspNet.Identity.Owin;
 using Microsoft.Owin.Security;
 using WebAppMVC.Models;
+using WebAppMVC.ViewModels;
 
 namespace WebAppMVC.Controllers
 {
