+++ WebAppMVC/ViewModels/ManageViewModels.cs
@@ -3,7 +3,7 @@ using System.ComponentModel.DataAnnotations;
 using Microsoft.AspNet.Identity;
 using Microsoft.Owin.Security;
 
-namespace WebAppMVC.Models
+namespace WebAppMVC.ViewModels
 {
     public class IndexViewModel
     {
