+++ WebAppMVC/WebAppMVC.csproj
@@ -235,16 +235,21 @@
     <Compile Include="Migrations\201709151221243_AddForeignKeyPropoertiesToPost.Designer.cs">
       <DependentUpon>201709151221243_AddForeignKeyPropoertiesToPost.cs</DependentUpon>
     </Compile>
+    <Compile Include="Migrations\201709171635483_AddNameToApplicationUser.cs" />
+    <Compile Include="Migrations\201709171635483_AddNameToApplicationUser.Designer.cs">
+      <DependentUpon>201709171635483_AddNameToApplicationUser.cs</DependentUpon>
+    </Compile>
     <Compile Include="Migrations\Configuration.cs" />
-    <Compile Include="Models\AccountViewModels.cs" />
+    <Compile Include="ViewModels\AccountViewModels.cs" />
     <Compile Include="Models\ApplicationDbContext.cs" />
     <Compile Include="Models\ApplicationUser.cs" />
     <Compile Include="Models\Categorie.cs" />
-    <Compile Include="Models\ManageViewModels.cs" />
+    <Compile Include="ViewModels\ManageViewModels.cs" />
     <Compile Include="Models\Post.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Startup.cs" />
     <Compile Include="ViewModels\PostFormViewModel.cs" />
+    <Compile Include="ViewModels\RegisterViewModel.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="App_Data\Nirvana.mdf" />
@@ -355,6 +360,9 @@
     </EmbeddedResource>
     <EmbeddedResource Include="Migrations\201709151221243_AddForeignKeyPropoertiesToPost.resx">
       <DependentUpon>201709151221243_AddForeignKeyPropoertiesToPost.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="Migrations\201709171635483_AddNameToApplicationUser.resx">
+      <DependentUpon>201709171635483_AddNameToApplicationUser.cs</DependentUpon>
     </EmbeddedResource>
   </ItemGroup>
   <PropertyGroup>
